Remove vestigial int32_t casts truncating file sizes in zmodem_recv_files()
Three comparisons between l and bytes cast bytes to int32_t, silently
truncating file sizes > 2GB. Both variables are int64_t — the casts
were left over from the 887147a69c "64-bitify" conversion where
long/ulong became int32_t/uint32_t before being later widened to
int64_t.

+++ src/sbbs3/zmodem.c
@@ -2041,13 +2041,13 @@ int zmodem_recv_files(zmodem_t* zm, const char* download_dir, uint64_t* bytes_re
 			if (fexist(fpath)) {
 				l = flength(fpath);
 				lprintf(zm, LOG_WARNING, "%s already exists (%" PRId64 " bytes)", fpath, l);
-				if (l >= (int32_t)bytes) {
+				if (l >= bytes) {
 					lprintf(zm, LOG_WARNING, "Local file size >= remote file size (%" PRId64 ")"
 					        , bytes);
 					if (zm->duplicate_filename == NULL)
 						break;
 					else {
-						if (l > (int32_t)bytes) {
+						if (l > bytes) {
 							if (zm->duplicate_filename(zm->cbdata, zm)) {
 								loop = TRUE;
 								continue;
@@ -2088,7 +2088,7 @@ int zmodem_recv_files(zmodem_t* zm, const char* download_dir, uint64_t* bytes_re
 					}
 					break;
 				}
-				if (l == (int32_t)bytes) {
+				if (l == bytes) {
 					lprintf(zm, LOG_INFO, "CRC, length, and filename match.");
 					break;
 				}
